refactor(web): fix lib->components layer inversion (closes #3220)
Two threads:

1. Delete `lib/useAuth.ts` (20-LOC wrapper that just forwarded
   `useSession()`). All ~10 callers now import `useSession` directly
   from `@/components/SessionProvider`. The returned shape is
   identical, so destructured fields stay unchanged. Removed the
   redundant `@/lib/useAuth` mock from `AccountSettings-heading-
   hierarchy.test.tsx` (the same file already mocked
   `@/components/SessionProvider` correctly), and rewrote
   `save-search-button.test.tsx` to mock `SessionProvider` instead.
   Deleted `useAuth.test.ts` (covered the wrapper that no longer
   exists; `SessionProvider.test.tsx` already covers the underlying
   context).

2. Move the `SelectedLocation` domain type out of the `"use client"`
   `components/search/location-pills.tsx` leaf into a new lib home
   at `lib/search/selected-location.ts`. `lib/actions/search-input.ts`
   (a `"use server"` server action) now imports from the new home,
   so the server-side type graph no longer reaches into a UI
   component. `location-pills.tsx` re-exports the type so the eight
   other (legitimately component-level) consumers compile unchanged
   without a wide rename.

Both inversions removed; no behavior change. 17 files touched
(2 deleted, 1 new, 14 edited).
