+++ apps/web/app/[lang]/(app)/watchlists/watchlists-page.tsx
@@ -5,7 +5,7 @@ import { useRouter, useSearchParams } from "next/navigation";
 import { Eye, Loader2, LogIn } from "lucide-react";
 import { Trans, useLingui } from "@lingui/react/macro";
 import { useLocalePath } from "@/lib/useLocalePath";
-import { useAuth } from "@/lib/useAuth";
+import { useSession } from "@/components/SessionProvider";
 import type { WatchlistSummary, WatchlistFilters } from "@/lib/actions/watchlists";
 import { createWatchlist } from "@/lib/actions/watchlists";
 import { WatchlistCard, CreateWatchlistCard } from "@/components/watchlist/watchlist-card";
@@ -25,7 +25,7 @@ export function WatchlistsPage({
   const { t } = useLingui();
   const router = useRouter();
   const lp = useLocalePath();
-  const { user, isLoggedIn } = useAuth();
+  const { user, isLoggedIn } = useSession();
   const searchParams = useSearchParams();
   const [creating, setCreating] = useState(false);
   const upgrade = useUpgradeModal();
