+++ apps/web/src/components/settings/__tests__/AccountSettings-heading-hierarchy.test.tsx
@@ -23,20 +23,11 @@ vi.mock("next/navigation", () => ({
   }),
 }));
 
-// SessionProvider drives useAuth — stub a logged-in user so the
-// component renders its real sections rather than the LoginPrompt
-// short-circuit.
+// AccountSettings reads its viewer via useSession() — stub a logged-in
+// user so the component renders its real sections rather than the
+// LoginPrompt short-circuit.
 vi.mock("@/components/SessionProvider", () => ({
   useSession: () => ({
-    user: { email: "test@example.com", username: "test" },
-    isLoggedIn: true,
-    isPending: false,
-    refresh: vi.fn(),
-  }),
-}));
-
-vi.mock("@/lib/useAuth", () => ({
-  useAuth: () => ({
     user: { email: "test@example.com", username: "test" },
     isLoggedIn: true,
     isPending: false,
